Header and refactors
