+++ app/codeSplitAssets/NyanCat.scss
@@ -10,12 +10,8 @@ $primary-color: #0098da;
 }
 
 .wrapper {
-  position: absolute;
-  top: 150px;
-  left: 50%;
   width: 400px;
   height: 300px;
-  margin-left: -200px;
   border-width: 1px;
   overflow: hidden;
   border-style: solid;
