+++ app/components/Page.jsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const Title = styled.h1`
+  font-size: var(--step-up-5);
+  margin-bottom: .5em;
+`
+
+export const Wrapper = styled.div`
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 1rem;
+  text-align: center;
+`
